feat. nav 상단탭으로 변경
