--- src/components/common/bottomTab.style.ts
@@ -1,34 +1,0 @@
-import { styled } from 'styled-components';
-
-export const Container = styled.div`
-  width: 100%;
-  height: 5.625rem;
-  position: absolute;
-  left: 0;
-  bottom: 0;
-  display: flex;
-  justify-content: space-between;
-  background-color: ${({ theme }) => theme.white};
-  box-shadow: ${({ theme }) => `0 -2px 6px 0 ${theme.shadow}`};
-  padding: 0 1.25rem;
-  text-align: center;
-  color: ${({ theme }) => theme.white};
-  button {
-    width: 5.234rem;
-    height: 3.5rem;
-    background: none;
-    font-size: ${({ theme }) => theme.fs12};
-    font-weight: ${({ theme }) => theme.fw700};
-    color: ${({ theme }) => theme.gray800};
-  }
-`;
-
-export const BadgeWrapper = styled.div`
-  position: relative;
-`;
-
-export const Badge = styled.img<{ $isMobile: boolean }>`
-  position: absolute;
-  top: 0;
-  right: ${({ $isMobile }) => ($isMobile ? '15px' : '20px')};
-`;
